+++ benchmarks/wavefront_convergence_capture.sh
@@ -283,6 +283,9 @@ def wavefront_metric_values(path)
     frontier_closest_hit_batch_rays: [],
     direct_light_any_hit_batch_chunks: [],
     direct_light_any_hit_batch_rays: [],
+    direct_light_any_hit_frontier_packed_ray_bytes: [],
+    direct_light_any_hit_frontier_host_query_bytes: [],
+    direct_light_any_hit_frontier_state_handle_bytes: [],
     frontier_query_round_trips: [],
     frontier_resident_query_round_trips_estimate: [],
     frontier_resident_query_round_trip_savings_estimate: [],
@@ -374,6 +377,9 @@ def wavefront_metric_values(path)
       frontier_closest_hit_batch_rays: 0.0,
       direct_light_any_hit_batch_chunks: 0.0,
       direct_light_any_hit_batch_rays: 0.0,
+      direct_light_any_hit_frontier_packed_ray_bytes: 0.0,
+      direct_light_any_hit_frontier_host_query_bytes: 0.0,
+      direct_light_any_hit_frontier_state_handle_bytes: 0.0,
       frontier_query_round_trips: 0.0,
       frontier_resident_query_round_trips_estimate: 0.0,
       frontier_resident_query_round_trip_savings_estimate: 0.0,
@@ -513,6 +519,12 @@ def wavefront_metric_values(path)
         batching.fetch("directLightAnyHitBatchChunksPerDepth", []).sum { |value| value.to_f }
       run_values[:direct_light_any_hit_batch_rays] +=
         batching.fetch("directLightAnyHitBatchRaysPerDepth", []).sum { |value| value.to_f }
+      run_values[:direct_light_any_hit_frontier_packed_ray_bytes] +=
+        batching.fetch("directLightAnyHitFrontierPackedRayBytes", 0).to_f
+      run_values[:direct_light_any_hit_frontier_host_query_bytes] +=
+        batching.fetch("directLightAnyHitFrontierHostQueryBytes", 0).to_f
+      run_values[:direct_light_any_hit_frontier_state_handle_bytes] +=
+        batching.fetch("directLightAnyHitFrontierStateHandleBytes", 0).to_f
       run_values[:frontier_query_round_trips] +=
         batching.fetch("frontierQueryRoundTrips", 0).to_f
       run_values[:frontier_resident_query_round_trips_estimate] +=
@@ -748,6 +760,9 @@ end
    frontier_closest_hit_batch_rays
    direct_light_any_hit_batch_chunks
    direct_light_any_hit_batch_rays
+   direct_light_any_hit_frontier_packed_ray_bytes
+   direct_light_any_hit_frontier_host_query_bytes
+   direct_light_any_hit_frontier_state_handle_bytes
    frontier_query_round_trips
    frontier_resident_query_round_trips_estimate
    frontier_resident_query_round_trip_savings_estimate
@@ -920,6 +935,9 @@ def aggregate_run(run)
     closest_hit_batch_rays: 0.0,
     any_hit_batch_chunks: 0.0,
     any_hit_batch_rays: 0.0,
+    direct_light_any_hit_frontier_packed_ray_bytes: 0.0,
+    direct_light_any_hit_frontier_host_query_bytes: 0.0,
+    direct_light_any_hit_frontier_state_handle_bytes: 0.0,
     frontier_round_trips: 0.0,
     resident_frontier_round_trips: 0.0,
     resident_frontier_savings: 0.0,
@@ -1008,6 +1026,12 @@ def aggregate_run(run)
     values[:closest_hit_batch_rays] += sum_array(batching, "frontierClosestHitBatchRaysPerDepth")
     values[:any_hit_batch_chunks] += sum_array(batching, "directLightAnyHitBatchChunksPerDepth")
     values[:any_hit_batch_rays] += sum_array(batching, "directLightAnyHitBatchRaysPerDepth")
+    values[:direct_light_any_hit_frontier_packed_ray_bytes] +=
+      batching.fetch("directLightAnyHitFrontierPackedRayBytes", 0).to_f
+    values[:direct_light_any_hit_frontier_host_query_bytes] +=
+      batching.fetch("directLightAnyHitFrontierHostQueryBytes", 0).to_f
+    values[:direct_light_any_hit_frontier_state_handle_bytes] +=
+      batching.fetch("directLightAnyHitFrontierStateHandleBytes", 0).to_f
     values[:frontier_round_trips] += batching.fetch("frontierQueryRoundTrips", 0).to_f
     values[:resident_frontier_round_trips] +=
       batching.fetch("frontierResidentQueryRoundTripsEstimate", 0).to_f
@@ -1152,7 +1176,7 @@ scene_dir = ARGV.fetch(0)
 queue_dirs = Dir.glob(File.join(scene_dir, "queue_*")).select { |path| File.directory?(path) }
 queue_dirs.sort_by! { |path| File.basename(path).delete_prefix("queue_").to_i }
 
-puts "queue_size variant render_ms primary_samples last_retained_active active_host_path_state_bytes tile_count tile_grid max_tile_width max_tile_height max_tile_pixels avg_tile_pixels avg_tile_samples max_tile_samples ray8_chunks ray4_chunks closest_hit_batch_chunks closest_hit_batch_rays any_hit_batch_chunks any_hit_batch_rays frontier_round_trips resident_frontier_round_trips resident_frontier_savings closest_hit_ray_upload_bytes any_hit_ray_upload_bytes closest_hit_query_transfer_bytes any_hit_query_transfer_bytes closest_hit_frontier_residency any_hit_frontier_residency closest_hit_frontier_packed_ray_bytes any_hit_frontier_packed_ray_bytes closest_hit_frontier_host_query_bytes any_hit_frontier_host_query_bytes closest_hit_frontier_state_handle_bytes any_hit_frontier_state_handle_bytes resident_frontiers_supported gpu_frontier_compaction_supported gpu_frontier_compaction_unavailable_reason prepared_ray_batch_compaction_supported resident_direct_light_batches_supported resident_direct_light_batches_unavailable_reason mixed_query_depths mixed_query_round_trips mixed_query_rays mixed_query_closest_hit_rays mixed_query_any_hit_rays packet_fill scalar_tail_fraction fallback_fraction scalar_rays fallback_rays frontier_compaction_passes frontier_compaction_input_samples frontier_compaction_retained_samples frontier_compaction_removed_samples frontier_compaction_removed_fraction frontier_compaction_moved_samples frontier_compaction_moved_retained_fraction frontier_compaction_retained_index_bytes frontier_compaction_input_host_path_state_bytes frontier_compaction_retained_host_path_state_bytes frontier_compaction_removed_host_path_state_bytes frontier_compaction_candidate_packed_ray_bytes frontier_compaction_candidate_state_handle_bytes frontier_compaction_candidate_host_path_state_bytes frontier_largest_compaction_candidate_packed_ray_bytes frontier_largest_compaction_candidate_state_handle_bytes frontier_largest_compaction_candidate_host_path_state_bytes compaction_execution sample_generation_worker_ms integrator_worker_ms integrator_frontier_partition_worker_ms integrator_residual_worker_ms"
+puts "queue_size variant render_ms primary_samples last_retained_active active_host_path_state_bytes tile_count tile_grid max_tile_width max_tile_height max_tile_pixels avg_tile_pixels avg_tile_samples max_tile_samples ray8_chunks ray4_chunks closest_hit_batch_chunks closest_hit_batch_rays any_hit_batch_chunks any_hit_batch_rays direct_light_any_hit_frontier_packed_ray_bytes direct_light_any_hit_frontier_host_query_bytes direct_light_any_hit_frontier_state_handle_bytes frontier_round_trips resident_frontier_round_trips resident_frontier_savings closest_hit_ray_upload_bytes any_hit_ray_upload_bytes closest_hit_query_transfer_bytes any_hit_query_transfer_bytes closest_hit_frontier_residency any_hit_frontier_residency closest_hit_frontier_packed_ray_bytes any_hit_frontier_packed_ray_bytes closest_hit_frontier_host_query_bytes any_hit_frontier_host_query_bytes closest_hit_frontier_state_handle_bytes any_hit_frontier_state_handle_bytes resident_frontiers_supported gpu_frontier_compaction_supported gpu_frontier_compaction_unavailable_reason prepared_ray_batch_compaction_supported resident_direct_light_batches_supported resident_direct_light_batches_unavailable_reason mixed_query_depths mixed_query_round_trips mixed_query_rays mixed_query_closest_hit_rays mixed_query_any_hit_rays packet_fill scalar_tail_fraction fallback_fraction scalar_rays fallback_rays frontier_compaction_passes frontier_compaction_input_samples frontier_compaction_retained_samples frontier_compaction_removed_samples frontier_compaction_removed_fraction frontier_compaction_moved_samples frontier_compaction_moved_retained_fraction frontier_compaction_retained_index_bytes frontier_compaction_input_host_path_state_bytes frontier_compaction_retained_host_path_state_bytes frontier_compaction_removed_host_path_state_bytes frontier_compaction_candidate_packed_ray_bytes frontier_compaction_candidate_state_handle_bytes frontier_compaction_candidate_host_path_state_bytes frontier_largest_compaction_candidate_packed_ray_bytes frontier_largest_compaction_candidate_state_handle_bytes frontier_largest_compaction_candidate_host_path_state_bytes compaction_execution sample_generation_worker_ms integrator_worker_ms integrator_frontier_partition_worker_ms integrator_residual_worker_ms"
 queue_dirs.each do |queue_dir|
   queue_size = File.basename(queue_dir).delete_prefix("queue_")
   Dir.glob(File.join(queue_dir, "wavefront_*.metrics.json")).sort.each do |metrics_path|
@@ -1205,6 +1229,9 @@ queue_dirs.each do |queue_dir|
       format("%.0f", median_for.call(:closest_hit_batch_rays)),
       format("%.0f", median_for.call(:any_hit_batch_chunks)),
       format("%.0f", median_for.call(:any_hit_batch_rays)),
+      format("%.0f", median_for.call(:direct_light_any_hit_frontier_packed_ray_bytes)),
+      format("%.0f", median_for.call(:direct_light_any_hit_frontier_host_query_bytes)),
+      format("%.0f", median_for.call(:direct_light_any_hit_frontier_state_handle_bytes)),
       format("%.0f", median_for.call(:frontier_round_trips)),
       format("%.0f", median_for.call(:resident_frontier_round_trips)),
       format("%.0f", median_for.call(:resident_frontier_savings)),